% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_path}
\alias{create_path}
\title{Title}
\usage{
create_path(path, type, directory)
}
\arguments{
\item{path}{a string path to the raw .R script}

\item{type}{string. type of path to return, either "run" or "bash"}

\item{directory}{the directory path}
}
\value{
a path to either a "_run.R" or "_bash.sh" file
}
\description{
Resolve the path in the same way that Slurm deals with the 'directory 'chdir' flag
https://slurm.schedmd.com/sbatch.html
\enumerate{
\item If directory starts with ".", then path is the current directory
\item If directory starts with a "/", then path is considered absolute and the file is located there
\item If directory starts with neither then a subdirectory is create below the raw .R script
}
}
