% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_argument}
\alias{get_argument}
\title{get_argument}
\usage{
get_argument(parameter_name)
}
\arguments{
\item{parameter_name}{the parameter to look for}
}
\value{
a string. argument passed
}
\description{
credit to https://gist.github.com/abelcallejo/f557a7d4ce7e37bbe5306dae864d0f0f#command-line
}
