% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_slurm.R
\name{here}
\alias{here}
\title{Here - where am I?}
\usage{
here()
}
\value{
file path to where \code{here()} was called.
}
\description{
https://gist.github.com/jasonsychau/ff6bc78a33bf3fd1c6bd4fa78bbf42e7
}
