% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_slurm.R
\name{queue}
\alias{queue}
\alias{queue,Slurm-method}
\title{queue}
\usage{
queue(object)

\S4method{queue}{Slurm}(object)
}
\arguments{
\item{object}{a Slurm object}
}
\value{
side effects, plus the Slurm object
}
\description{
A short description...
}
