% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_slurm.R
\name{write_preamble}
\alias{write_preamble}
\alias{write_preamble,Slurm-method}
\title{write_preamble}
\usage{
write_preamble(object)

\S4method{write_preamble}{Slurm}(object)
}
\arguments{
\item{object}{a Slurm object}
}
\value{
a string, the Slurm preamble
}
\description{
A short description...
}
