% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_slurm.R
\name{write_r_script}
\alias{write_r_script}
\alias{write_r_script,Slurm-method}
\title{write_r_script}
\usage{
write_r_script(object)

\S4method{write_r_script}{Slurm}(object)
}
\arguments{
\item{object}{a Slurm object}
}
\value{
a string, the r script
}
\description{
A short description...
}
