% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_slurm.R
\docType{class}
\name{Slurm-class}
\alias{Slurm-class}
\alias{Slurm}
\title{Slurm}
\value{
a Slurm object
}
\description{
TODO: the description
}
\section{Slots}{

\describe{
\item{\code{job_name}}{character.}

\item{\code{account}}{character.}

\item{\code{partition}}{character.}

\item{\code{nodes}}{integer}

\item{\code{tasks_per_node}}{integer}

\item{\code{cpu_per_task}}{integer}

\item{\code{max_time_days}}{integer}

\item{\code{max_time_hours}}{integer}

\item{\code{max_time_mins}}{integer}

\item{\code{max_time_secs}}{integer}

\item{\code{mem}}{integer.}

\item{\code{mem_unit}}{character.}

\item{\code{directory}}{character.}

\item{\code{modules}}{character.}

\item{\code{r_version}}{character.}

\item{\code{rscript_options}}{character.}

\item{\code{array}}{integer array c(1,2,3,4) etc}
}}

